Update.
2002-11-24  Robert Love  <rml@tech9.net>

	* posix/sched.h: Second parameter of sched_setaffinity and
	sched_getaffinity is `unsigned int', not `unsigned long'.
	* sysdeps/generic/sched_setaffinity.c: Likewise.
	* sysdeps/generic/sched_getaffinity.c: Likewise.

	Reported by John Levon <levon@movementarian.org>.

+++ sysdeps/generic/sched_getaffinity.c
@@ -1,4 +1,4 @@
-/* Copyright (C) 2002 Free Software Foundation, Inc.
+/* Copyright (C) 2002, 2003 Free Software Foundation, Inc.
    This file is part of the GNU C Library.
 
    The GNU C Library is free software; you can redistribute it and/or
@@ -25,7 +25,7 @@
 int
 sched_getaffinity (pid, len, mask)
      pid_t pid;
-     unsigned long int len;
+     unsigned int len;
      unsigned long int *mask;
 {
   __set_errno (ENOSYS);
